Move AppTagInfo to new module
