+++ nexus-common/src/universal_tag/mod.rs
@@ -1,1 +1,2 @@
+pub mod app_tag_info;
 pub mod homeserver_parsed_uri;
